+++ src/ray/raylet/scheduling/cluster_resource_scheduler.cc
@@ -1186,3 +1186,7 @@ void ClusterResourceScheduler::FreeLocalTaskResources(
   FreeTaskResourceInstances(task_allocation);
   UpdateLocalAvailableResourcesFromResourceInstances();
 }
+
+// void ClusterResourceScheduler::Heartbeat(std::shared_ptr<HeartbeatTableData> heartbeat_data) {
+
+// }
