+++ src/ray/raylet/scheduling/cluster_resource_scheduler.h
@@ -21,160 +21,15 @@
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/flat_hash_set.h"
 #include "ray/common/task/scheduling_resources.h"
+#include "ray/gcs/pb_util.h"
 #include "ray/raylet/scheduling/fixed_point.h"
+#include "ray/raylet/scheduling/cluster_resource_helpers.h"
 #include "ray/raylet/scheduling/scheduling_ids.h"
 #include "ray/util/logging.h"
 
-/// List of predefined resources.
-enum PredefinedResources { CPU, MEM, GPU, TPU, PredefinedResources_MAX };
 // Specify resources that consists of unit-size instances.
 static std::unordered_set<int64_t> UnitInstanceResources{CPU, GPU, TPU};
 
-/// Helper function to compare two vectors with FixedPoint values.
-bool EqualVectors(const std::vector<FixedPoint> &v1, const std::vector<FixedPoint> &v2);
-
-/// Convert a vector of doubles to a vector of resource units.
-std::vector<FixedPoint> VectorDoubleToVectorFixedPoint(const std::vector<double> &vector);
-
-/// Convert a vector of resource units to a vector of doubles.
-std::vector<double> VectorFixedPointToVectorDouble(
-    const std::vector<FixedPoint> &vector_fp);
-
-struct ResourceCapacity {
-  FixedPoint total;
-  FixedPoint available;
-};
-
-/// Capacities of each instance of a resource.
-struct ResourceInstanceCapacities {
-  std::vector<FixedPoint> total;
-  std::vector<FixedPoint> available;
-};
-
-struct ResourceRequest {
-  /// Amount of resource being requested.
-  FixedPoint demand;
-  /// Specify whether the request is soft or hard.
-  /// If hard, the entire request is denied if the demand exceeds the resource
-  /// availability. Otherwise, the request can be still be granted.
-  /// Prefernces are given to the nodes with the lowest number of violations.
-  bool soft;
-};
-
-/// Resource request, including resource ID. This is used for custom resources.
-struct ResourceRequestWithId : ResourceRequest {
-  /// Resource ID.
-  int64_t id;
-};
-
-// Data structure specifying the capacity of each resource requested by a task.
-class TaskRequest {
- public:
-  /// List of predefined resources required by the task.
-  std::vector<ResourceRequest> predefined_resources;
-  /// List of custom resources required by the task.
-  std::vector<ResourceRequestWithId> custom_resources;
-  /// List of placement hints. A placement hint is a node on which
-  /// we desire to run this task. This is a soft constraint in that
-  /// the task will run on a different node in the cluster, if none of the
-  /// nodes in this list can schedule this task.
-  absl::flat_hash_set<int64_t> placement_hints;
-  /// Returns human-readable string for this task request.
-  std::string DebugString() const;
-};
-
-// Data structure specifying the capacity of each instance of each resource
-// allocated to a task.
-class TaskResourceInstances {
- public:
-  /// The list of instances of each predifined resource allocated to a task.
-  std::vector<std::vector<FixedPoint>> predefined_resources;
-  /// The list of instances of each custom resource allocated to a task.
-  absl::flat_hash_map<int64_t, std::vector<FixedPoint>> custom_resources;
-  bool operator==(const TaskResourceInstances &other);
-  /// For each resource of this request aggregate its instances.
-  TaskRequest ToTaskRequest() const;
-  /// Get CPU instances only.
-  std::vector<FixedPoint> GetCPUInstances() const {
-    if (!this->predefined_resources.empty()) {
-      return this->predefined_resources[CPU];
-    } else {
-      return {};
-    }
-  };
-  std::vector<double> GetCPUInstancesDouble() const {
-    if (!this->predefined_resources.empty()) {
-      return VectorFixedPointToVectorDouble(this->predefined_resources[CPU]);
-    } else {
-      return {};
-    }
-  };
-  /// Get GPU instances only.
-  std::vector<FixedPoint> GetGPUInstances() const {
-    if (!this->predefined_resources.empty()) {
-      return this->predefined_resources[GPU];
-    } else {
-      return {};
-    }
-  };
-  std::vector<double> GetGPUInstancesDouble() const {
-    if (!this->predefined_resources.empty()) {
-      return VectorFixedPointToVectorDouble(this->predefined_resources[GPU]);
-    } else {
-      return {};
-    }
-  };
-  /// Get mem instances only.
-  std::vector<FixedPoint> GetMemInstances() const {
-    if (!this->predefined_resources.empty()) {
-      return this->predefined_resources[MEM];
-    } else {
-      return {};
-    }
-  };
-  std::vector<double> GetMemInstancesDouble() const {
-    if (!this->predefined_resources.empty()) {
-      return VectorFixedPointToVectorDouble(this->predefined_resources[MEM]);
-    } else {
-      return {};
-    }
-  };
-  /// Check whether there are no resource instances.
-  bool IsEmpty() const;
-  /// Returns human-readable string for these resources.
-  std::string DebugString() const;
-};
-
-/// Total and available capacities of each resource of a node.
-class NodeResources {
- public:
-  /// Available and total capacities for predefined resources.
-  std::vector<ResourceCapacity> predefined_resources;
-  /// Map containing custom resources. The key of each entry represents the
-  /// custom resource ID.
-  absl::flat_hash_map<int64_t, ResourceCapacity> custom_resources;
-  /// Returns if this equals another node resources.
-  bool operator==(const NodeResources &other);
-  /// Returns human-readable string for these resources.
-  std::string DebugString(StringIdMap string_to_int_map) const;
-};
-
-/// Total and available capacities of each resource instance.
-/// This is used to describe the resources of the local node.
-class NodeResourceInstances {
- public:
-  /// Available and total capacities for each instance of a predefined resource.
-  std::vector<ResourceInstanceCapacities> predefined_resources;
-  /// Map containing custom resources. The key of each entry represents the
-  /// custom resource ID.
-  absl::flat_hash_map<int64_t, ResourceInstanceCapacities> custom_resources;
-  /// Extract available resource instances.
-  TaskResourceInstances GetAvailableResourceInstances();
-  /// Returns if this equals another node resources.
-  bool operator==(const NodeResourceInstances &other);
-  /// Returns human-readable string for these resources.
-  std::string DebugString(StringIdMap string_to_int_map) const;
-};
 
 /// Class encapsulating the cluster resources and the logic to assign
 /// tasks to nodes based on the task's constraints and the available
@@ -217,6 +72,10 @@ class ClusterResourceScheduler {
   /// False otherwise.
   bool AllocateTaskResources(int64_t node_id, const TaskRequest &task_req,
                              std::shared_ptr<TaskResourceInstances> task_allocation);
+
+  /// The previous heartbeat. Necessary for the implementation of light
+  /// heartbeats.
+  /* HeartbeatTableData prev_heartbeat_; */
 
  public:
   ClusterResourceScheduler(void){};
@@ -512,6 +371,11 @@ class ClusterResourceScheduler {
   /// following availabilities: 0.2, 0.3, 0.1, 1. Then the total GPU
   // resources availabile at that node is 0.2 + 0.3 + 0.1 + 1. = 1.6
   void UpdateLocalAvailableResourcesFromResourceInstances();
+
+  /// Fill out the HeartbeatTableData. This is primarily used for raylet <->
+  /// GCS heartbeats. Specifically, this fills out the resources_available and
+  /// resources_total fields.
+  /* void Heartbeat(std::shared_ptr<HeartbeatTableData> heartbeat_data); */
 
   /// Return human-readable string for this scheduler state.
   std::string DebugString() const;
